fix: extend RestaurantFetch interface into two childen
Previously, the RestaurantFetch had two nullable methods, fetchRestaurants
and fetchRecommendations, but fetchRestaurants was triggering a TS error
as it was possible undefined (that is, nullable). To get around this,
the interfaces SearchRestaurantFetch and RecommendedRestaurantFetch
both now extend from RestaurantFetch and have one and only one of the
methods mentioned above.